Override spliterator method
The persistent collection types relied on default implementations of the
spliterator method coming from List and Set in many places, which meant
that some important spliterator characteristics were missing. For example,
all of the collections are immutable and have a defined iteration order,
but that was not communicated in the Spliterator characteristics.